+++ R/globals.R
@@ -1,6 +1,7 @@
 # This tells R CMD check that these variables are used intentionally
 # inside dplyr/ggplot2 pipes and are not global variables.
 utils::globalVariables(c(
+  ".SD",
   ".time",
   ".treatment",
   "id",